Refactor utils unit tests.
